General improvement of the  Web Client

+++ blue-web/src/main/resources/webapp/js/paper/controllers/PapersController.js
@@ -1,28 +1,51 @@
+/*
+ * This file is part of the \BlueLaTeX project.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Paper.Services.Paper'])
-  .controller('PapersController', ['$rootScope', '$scope', 'PaperService','$log','MessagesService','$localStorage',
-    function ($rootScope, $scope, PaperService,$log,MessagesService,$localStorage) {
+  .controller('PapersController', ['$rootScope', '$scope', 'PaperService','$log','MessagesService','$localStorage','localize',
+    function ($rootScope, $scope, PaperService,$log,MessagesService,$localStorage,localize) {
+      // list of peper
       $scope.papers = [];
 
-      if($localStorage.userPaperReverse == null)
+      // load settings from localStorage
+      $scope.reverse = $localStorage.userPaperReverse;
+      if($scope.reverse == null)
         $scope.reverse = false;
-      else
-        $scope.reverse = $localStorage.userPaperReverse;
+
       $scope.$watch("reverse", function(value) {
         $localStorage.reverse = value;
       });
 
-      if($localStorage.userPaperPredicate == null)
+      $scope.predicate = $localStorage.userPaperPredicate;
+      if($scope.predicate == null)
         $scope.predicate = 'title';
-      else
-        $scope.predicate = $localStorage.userPaperPredicate;
+        
       $scope.$watch("predicate", function(value) {
         $localStorage.userPaperPredicate = value;
       });
 
-      if($localStorage.userPaperStyle == null)
+      $scope.orderBy = function(predicate, reverse) {
+        $scope.predicate = predicate;
+        $scope.reverse = reverse;
+      };
+
+      $scope.display_style = $localStorage.userPaperStyle;
+      if($scope.display_style == null)
         $scope.display_style = 'list';
-      else
-        $scope.display_style = $localStorage.userPaperStyle;
+
       $scope.$watch("display_style", function(value) {
         $localStorage.userPaperStyle = value;
       });
@@ -31,13 +54,9 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
       $scope.role_filter = 'all';
       $scope.tag_filter = 'all';
 
+      // get paper of a user
       PaperService.getUserPapers($rootScope.loggedUser).then(function (data) {
-        $scope.papers = [];
-        for (var i = 0; i < data.length; i++) {
-          data[i].date = new Date();
-
-          $scope.papers.push(data[i]);
-        }
+        $scope.papers = data;
       }, function (err) {
         MessagesService.clear();
         switch (err.status) {
@@ -52,6 +71,9 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
         }
       });
 
+      /***************/
+      /* Date Filter */
+      /***************/
       var dateFilterToday = function (paper) {
         var now = new Date();
         return paper.date.getDate() == now.getDate() &&
@@ -59,6 +81,7 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
           paper.date.getFullYear() == now.getFullYear();
       };
       $scope.dateFilterToday = dateFilterToday;
+
       var dateFilterYesterday = function (paper) {
         var yesterday = new Date();
         yesterday.setDate(yesterday.getDate() - 1);
@@ -67,17 +90,20 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
           paper.date.getFullYear() == yesterday.getFullYear();
       };
       $scope.dateFilterYesterday = dateFilterYesterday;
+
       var dateFilterLWeek = function (paper) {
         var now = new Date();
         return paper.date.getWeek() == now.getWeek() - 1;
       };
       $scope.dateFilterLWeek = dateFilterLWeek;
+
       var dateFilterTWeek = function (paper) {
         var now = new Date();
         return paper.date.getWeek() == now.getWeek() &&
           paper.date.getFullYear() == now.getFullYear();
       };
       $scope.dateFilterTWeek = dateFilterTWeek;
+
       var dateFilterMonth = function (paper) {
         var now = new Date();
         return paper.date.getMonth() == now.getMonth() &&
@@ -109,6 +135,9 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
         }
       };
 
+      /***************/
+      /* Role filter */
+      /***************/
       var roleFilterAuthor = function (paper) {
         return paper.role == 'author';
       };
@@ -130,7 +159,11 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
         }
       };
 
+      /**
+      * Delete a paper
+      */
       $scope.delete = function (paper) {
+        if(!confirm(localize.getLocalizedString('_Delete_paper_confirm_', paper.title))) return;
         PaperService.delete(paper.id).then(function (data) {
           if (data.response == true) {
             $scope.papers.splice($scope.papers.indexOf(paper), 1);
@@ -155,7 +188,7 @@ angular.module('bluelatex.Paper.Controllers.Papers', ['ngStorage','bluelatex.Pap
       };
 
       //action listener: action in the menu
-      $scope.$on('handleAction', function (event, data) {
+      $rootScope.$on('handleMenuAction', function (event, data) {
         if ($scope[data]) {
           $scope[data]();
         }
